pyproject.toml
[tool.poetry]
name = "django_template"
version = "0.1.0"
description = "Template Django project for JS+ back-end applications"
authors = ["Juntos Somos Mais <labs@juntossomosmais.com.br>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4"
djangorestframework = "^3"
psycopg2-binary = "^2"
### Infrastructure
django-health-check = "^3"
django-cors-headers = "^3"
### WSGI
gunicorn = "^20"
gevent = "^21"
### Logging
request-id-django-log = "^0"
python-json-logger = "^2"
### Tooling
requests = "^2"
### Helpers and so on
autodynatrace = "^1"

[tool.poetry.dev-dependencies]
### Pytest: Essentials
pytest = "^7"
pytest-mock = "^3"
pytest-django = "^4"
pytest-cov = "^3"
pytest-env = "^0"
pytest-xdist = "*"
### Pytest: Add-ons
pytest-icdiff = "^0"
pytest-clarity = "^1"
### Format, lint, static type checker, among others
black = "*"
mypy = "^0"
isort = "*"
flake8 = "*"
flake8-bugbear = "*"
autoflake = "^1"
pre-commit = "^2"
types-requests = "*"

[tool.black]
line-length = 120
target_version = ["py310"]
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | settings.py
  )/
)
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
